requirements.txt
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.0.1
django-webpack-loader==0.5.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
model-mommy==1.5.0
oauthlib==2.0.6
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Unipath==1.1
urllib3==1.22
wcwidth==0.1.7
